Move ConfigReader::ConnectionConfig to Sysconfig
