+++ test/y2network/sysconfig/connection_config_reader_test.rb
@@ -17,13 +17,13 @@
 # To contact SUSE LLC about this file by physical or electronic mail, you may
 # find current contact information at www.suse.com.
 
-require_relative "../../../test_helper"
+require_relative "../../test_helper"
 
-require "y2network/config_reader/connection_config/sysconfig"
-require "y2network/config_reader/connection_config/sysconfig_handlers/wlan"
+require "y2network/sysconfig/connection_config_reader"
+require "y2network/sysconfig/connection_config_reader_handlers/wlan"
 require "y2network/physical_interface"
 
-describe Y2Network::ConfigReader::ConnectionConfig::Sysconfig do
+describe Y2Network::Sysconfig::ConnectionConfigReader do
   subject(:reader) { described_class.new }
 
   describe "#read" do
@@ -31,20 +31,20 @@ describe Y2Network::ConfigReader::ConnectionConfig::Sysconfig do
     let(:connection_config) { double("connection_config") }
     let(:handler) do
       instance_double(
-        Y2Network::ConfigReader::ConnectionConfig::SysconfigHandlers::Wlan,
+        Y2Network::Sysconfig::ConnectionConfigReaderHandlers::Wlan,
         connection_config: connection_config
       )
     end
 
     before do
       allow(reader).to receive(:require).and_call_original
-      allow(Y2Network::ConfigReader::ConnectionConfig::SysconfigHandlers::Wlan)
+      allow(Y2Network::Sysconfig::ConnectionConfigReaderHandlers::Wlan)
         .to receive(:new).and_return(handler)
     end
 
     it "uses the appropiate handler" do
       expect(reader).to receive(:require)
-        .with("y2network/config_reader/connection_config/sysconfig_handlers/wlan")
+        .with("y2network/sysconfig/connection_config_reader_handlers/wlan")
       conn = reader.read(interface, :wlan)
       expect(conn).to be(connection_config)
     end
